sprint_14
The Tracker application model has been created

1. The following controllers were added:

a. TrackersViewController - The main controller;
b. CreatingTrackerViewController - The transition controller for creating "Habits" and "Irregular events";
c. CreatingHabitViewController - The controller for creating Habits" and "Irregular events";
d. CategoryViewController - The controller for selecting "Categories";
e. NewCategoryViewController - The controller to create a new "Category";
f. ScheduleViewController - The controller for selecting the "Schedule".

2. The creation of cells has been implemented:

a. CustomTableViewCell - Shows the contents of the table on the "CreatingHabitViewController" and "ScheduleViewController" screens;
b. TrackerCollectionViewCell - Shows the collection cell on the "TrackersViewController" screen;
c. SupplementaryView - Shows the title of the created cell "TrackerCollectionViewCell" in "TrackersViewController".

3. The following entities have been created:

a. struct Tracker is an entity for storing information about a tracker (for a "Habit" or "Irregular Event");
b. struct TrackerCategory is an entity for storing trackers by category.
c. struct TrackerRecord is an entity for storing a record that a certain tracker was completed on a certain date;
d. enum TypeTrackers is an entity for determining a "Habit" or "Irregular Event";
e. enum WeekDay is an entity for creating a "Schedule".

4. Added a class that stores static data:

a. DataManager - Contains a ready-made list of "Categories" and "Trackers".